Implement folding of free-form geometry body
