+++ src/main/resources/META-INF/plugin.xml
@@ -102,7 +102,7 @@
                 implementationClass="it.czerwinski.intellij.wavefront.lang.ObjRefactoringSupportProvider"/>
         <lang.foldingBuilder
                 language="OBJ"
-                implementationClass="it.czerwinski.intellij.wavefront.lang.ObjFoldingBuilder"/>
+                implementationClass="it.czerwinski.intellij.wavefront.lang.folding.ObjFoldingBuilder"/>
         <lang.psiStructureViewFactory
                 language="OBJ"
                 implementationClass="it.czerwinski.intellij.wavefront.lang.structure.ObjStructureViewFactory"/>
@@ -191,7 +191,7 @@
                 implementationClass="it.czerwinski.intellij.wavefront.lang.MtlFindUsagesProvider"/>
         <lang.foldingBuilder
                 language="MTL"
-                implementationClass="it.czerwinski.intellij.wavefront.lang.MtlFoldingBuilder"/>
+                implementationClass="it.czerwinski.intellij.wavefront.lang.folding.MtlFoldingBuilder"/>
         <gotoSymbolContributor
                 implementation="it.czerwinski.intellij.wavefront.lang.MtlChooseByNameContributor"/>
         <lang.psiStructureViewFactory
